로그인 된 사용자만 댓글 작성 및 삭제 허용

+++ articles/views.py
@@ -78,19 +78,24 @@ def delete(request, pk):
 
 @require_POST
 def comments_create(request,pk) :
-    article = Article.objects.get(pk=pk)
-    comment_form = CommentForm(request.POST)
-    if comment_form.is_valid() :
-        # commit의 기본값은 true이다. commit을 False로 바꿔서 DB에 저장은 안하고 인스턴스는 만들어줌 
-        comment = comment_form.save(commit=False)
-        # 누락된 article입력 해주고 
-        comment.article = article
-        #세이브 해줌 
-        comment.save()
-    return redirect('articles:detail',article.pk)
+    if request.user.is_authenticated :
+        article = get_object_or_404(Article,pk=pk)
+        comment_form = CommentForm(request.POST)
+        if comment_form.is_valid() :
+            # commit의 기본값은 true이다. commit을 False로 바꿔서 DB에 저장은 안하고 인스턴스는 만들어줌 
+            comment = comment_form.save(commit=False)
+            # 누락된 article입력 해주고 
+            comment.article = article
+            #세이브 해줌 
+            comment.save()
+        return redirect('articles:detail',article.pk)
+    return redirect('accounts:login')
 
+@require_POST
 def comments_delete(request,article_pk,comment_pk) :
-    comment = Comment.objects.get(pk=comment_pk)
-    # article_pk = comment.article.pk
-    comment.delete()
-    return redirect('articles:detail',article_pk)
+    if request.user.is_authenticated :
+        comment = get_object_or_404(Comment,pk=comment_pk)
+        # article_pk = comment.article.pk
+        comment.delete()
+        return redirect('articles:detail',article_pk)
+    return redirect('accounts:login')
